requirement.txt
boto3
requests
begins
